Create generic header file
